Rename SuccessorsMixin to SuccessorsEngine, move to successors module
